Add fatalError() to general namespace
